okay

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i = 0;
-	long j = i, k = 2;
+	int j = i, k = 2;
 
 	while (i < 50)
 	{
